BasicMD.jl Documentation

Available Samplers

The desired sampler structure must be initialized before any sampling can be performed.

Metropolis Samplers

BasicMD.RWM — Method

RWM(V, β, Δt)

Set up the RWM sampler for Boltzmann.

Fields

V - Potential

β - Inverse temperature

Δt - Time step

BasicMD.MALA — Method

MALA(V, ∇V!, β, Δt)

Set up the MALA sampler for overdamped Langevin.

Fields

V - Potential

∇V! - In place gradient of the potential

β - Inverse temperature

Δt - Time step

BasicMD.HMC — Method

HMC(V, ∇V!, β, M, Δt, nΔt)

Set up the HMC sampler for Boltzmann.

Fields

V - Potential

∇V! - In place gradient of the potential

β - Inverse temperature

Δt - Time step

nΔt - Number of time steps to use in each Verlet run

Non-Metropolis Samplers

BasicMD.EM — Method

EM(∇V!, β, γ, M, Δt)

Set up the EM integrator for overdamped Langevin.

Fields

∇V! - In place gradient of the potential

β - Inverse temperature

Δt - Time step

BasicMD.ABOBA — Method

ABOBA(∇V!, β, γ, M, Δt)

Set up the ABOBA integrator for inertial Langevin.

Fields

∇V! - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

BasicMD.BAOAB — Method

BAOAB(∇V!, β, γ, M, Δt)

Set up the BAOAB integrator for inertial Langevin.

Fields

∇V! - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

BasicMD.BBK — Method

BBK(∇V!, β, γ, M, Δt)

Set up the BBK integrator for inertial Langevin.

Fields

∇V! - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

BasicMD.GJF — Method

GJF(∇V!, β, γ, M, Δt)

Set up the G-JF integrator for inertial Langevin.

Fields

∇V! - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

Options

These options set the number of iterations and the frequency at which data is recorderd.

BasicMD.MDOptions — Method

MDOptions(;n_iters = 10^4, n_save_iters=1)

Set options for samplers.

Fields

n_iters - Set the number of iterations of the sampler

nsaveiters - Set the frequency at which iterations are saved. If nsaveiters=1, every iteration is saved. If nsaveiters=n_iters, only the final iteration is saved.

Running Samplers