Non-Metropolis Samplers

These methods do not include a Metropolis-Hastings step, and, consequently, will sample from a distribution, $\mu_{\Delta t}(x)$, which is a biased approximation of $\mu(x) \propto e^{-\beta V(x)}$. This bias vanishes with Δt, and is often negligible in comparison to the statistical variance error.

First Order Methods

These methods are in the spirit of first order in time discretizations.

BasicMD.EM — Method

EM(∇V!, β, Δt)

Set up the EM integrator for overdamped Langevin.

Fields

∇V! - In place gradient of the potential

β - Inverse temperature

Δt - Time step

Second Order Methods

These methods are in the spirit of second order in time discretizations.

BasicMD.ABOBA — Method

ABOBA(∇V!, β, γ, M, Δt)

Set up the ABOBA integrator for inertial Langevin.

Fields

∇V! - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

BasicMD.BAOAB — Method

BAOAB(∇V!, β, γ, M, Δt)

Set up the BAOAB integrator for inertial Langevin.

Fields

∇V! - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

BasicMD.BBK — Method

BBK(∇V!, β, γ, M, Δt)

Set up the BBK integrator for inertial Langevin.

Fields

∇V! - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

BasicMD.GJF — Method

GJF(∇V!, β, γ, M, Δt)

Set up the G-JF integrator for inertial Langevin.

Fields

∇V! - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step