Options and Auxiliary Functions

Sampler Options

These options set the number of iterations and the frequency at which data is recorderd. This is generically used by all samplers.

BasicMD.MDOptions — Method

MDOptions(;n_iters = 10^4, n_save_iters=1)

Set options for samplers.

Fields

n_iters - Set the number of iterations of the sampler

nsaveiters - Set the frequency at which iterations are saved. If nsaveiters=1, every iteration is saved. If nsaveiters=n_iters, only the final iteration is saved.

Verlet Integrator

While the main goal of this package is to sample from NVT type ensembles, as it is needed for HMC, the Verlet integrator is included should one wish to sample from the NVE ensemble:

BasicMD.Verlet — Type

Verlet(∇V!, M, Δt)

Set up the Verlet integrator.

Fields

∇V! - In place gradient of the potential

M - Mass (either scalar or vector)

Δt - Time step